new update

+++ controllers/course/course.js
@@ -1,6 +1,7 @@
 import Course from "../../models/course/course.js";
 import User from "../../models/user/user.js";
 import ApiError from "../../middleware/errors/customError.js";
+import CourseSection from "../../models/course/course-section.js";
 export const createNewCourse = async (req, res, next) => {
    if(!req.isInstructor) return next(new ApiError("You are not authorized to create a course", 403));
   const NewCourse = new Course({ ...req.body  , instructorID: req.userId });
@@ -23,6 +24,9 @@ export const getCourseById = async (req, res, next) => {
   const { courseId } = req.params;
   try {
     const course = await Course.findById(courseId);
+    if(!course) return next(new ApiError("Course not found", 404));
+    const courseSections = await CourseSection.find({courseID: courseId});
+    course.courseSections = courseSections;
     res.status(200).json(course);
   } catch (error) {
     next(error);
@@ -43,6 +47,8 @@ export const updateCourse = async (req, res, next) => {
   }
 };
 export const deleteCourse = async (req, res, next) => {
+  if(!req.isInstructor) return next(new ApiError("You are not authorized to update a course", 403));
+  if(req.userId !== req.body.instructorID) return next(new ApiError("You are not authorized to update this course", 403));
   const { courseId } = req.params;
   try {
     await Course.findByIdAndDelete(courseId);
